+++ package.json
@@ -36,13 +36,13 @@
     "foundation-emails": "^2.3.1",
     "identity-style-guide": "^6.4.1",
     "intl-tel-input": "^17.0.8",
-    "libphonenumber-js": "^1.9.52",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "source-map-loader": "^3.0.1",
     "webpack": "^5.65.0",
     "webpack-assets-manifest": "^5.0.6",
     "webpack-cli": "^4.9.1",
+    "yarn-deduplicate": "^5.0.0",
     "zxcvbn": "4.4.2"
   },
   "devDependencies": {
